\name{insuranceDataGraph-package}
\alias{insuranceDataGraph-package}
\alias{insuranceDataGraph}
\docType{package}
\title{
\packageTitle{insuranceDataGraph}
}
\description{
\packageDescription{insuranceDataGraph}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{insuranceDataGraph}
\packageIndices{insuranceDataGraph}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{insuranceDataGraph}

Maintainer: \packageMaintainer{insuranceDataGraph}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
